feat(state): moved creators and getter in  helpers in state
