+++ packages/comments-react-components/src/state/Comments.js
@@ -1,7 +1,6 @@
 import { isInit, isFetching, isUpdating } from './selectors'
 import { CommentsStateError, UPDATE_IN_PROGRESS_ERROR } from './errors'
 import {
-  getDefaultState,
   initialize,
   initializeFail,
   initializeSuccess,
@@ -13,7 +12,9 @@ import {
   fetching
 } from './reducers'
 import { removeCommentFromResource, setCommentToResource } from './reducers/resource'
+import { getDefaultState } from './helpers/getters'
 
+console.log(12345)
 export const STATE_FIELD_NAME = 'commentsState'
 
 export class CommentsState {
